refactor: extract SessionDao methods from ActivityPackageDao in a separete SessionDao
