+++ app/src/main/java/edu/monash/student/happyactive/data/ActivityPackageDao.java
@@ -56,32 +56,4 @@ public abstract class ActivityPackageDao {
         }
         insertTaskList(tasks);
     }
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    public abstract long insertSession(ActivitySession activitySession);
-
-    @Query("SELECT * FROM activitySession")
-    public abstract LiveData<List<ActivitySession>> getAllSessionsRecords();
-
-    @Transaction
-    @Query("SELECT * FROM activityPackage WHERE id = :id")
-    public abstract LiveData<List<ActivityWithSessions>> getActivityWithSessionById(long id);
-
-    public void insertActivityWithSession(ActivityPackage activityPackage, ActivitySession session){
-        long id = insertActivity(activityPackage);
-        session.activityId = id;
-        insertSession(session);
-    }
-
-    public  void cancelSession(ActivitySession session){
-        session.completedDateTime = new Date();
-        session.status = ActivityPackageStatus.CANCELED;
-        updateSession(session);
-    }
-
-    @Update(onConflict = OnConflictStrategy.REPLACE)
-    public abstract void updateSession(ActivitySession activitySession);
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    public abstract void addNewJournalEntry(ActivityJournal journalEntry);
 }
